# server/api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "HARDWARESTORE-APP(REST API)",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "localhost"
    }
  ],
...
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/productimages/update/1": {
      "put": {
        "tags": [
          "product images"
        ],
        "summary": "Update images",
        "operationId": "updateimages",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "access_token",
            "in": "header",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "file",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "primary",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Status Has Been Update"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "36"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"24-tvVN/AbBWKY/dM4qwYyrqLLr33w\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 19:45:43 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/updateimagessuccess"
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Token Not Found"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "29"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1d-5HbAy3vDHKFvxHIXreOKe2CeD5o\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 19:46:25 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/updateimagesfail1"
            }
          }
        },
        "deprecated": false
      }
    },
    "/productimages/delete/3": {
      "delete": {
        "tags": [
          "product images"
        ],
        "summary": "Delete images",
        "operationId": "deleteimages",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "primary",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": 1
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "1"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1-NWoZK3kTsExUV00Ywo1G5jlUKKs\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 19:47:02 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Token Not Found"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "29"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1d-5HbAy3vDHKFvxHIXreOKe2CeD5o\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 19:47:27 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/deleteimagesfail1"
            }
          }
        },
        "deprecated": false
      }
    },
...
    "/line_items/add": {
      "post": {
        "tags": [
          "line item"
        ],
        "summary": "Add Line item",
        "operationId": "AddLineitem",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "access_token",
            "in": "header",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "qty",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "examples": {
              "application/json; charset=utf-8": {
                "id": 1,
                "qty": 2,
                "status": "Close",
                "updatedAt": "2021-08-17T13:15:14.657Z",
                "createdAt": "2021-08-17T13:15:14.657Z"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "165"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"a5-UXf6cAAnPXf1bvrBIU+sLS8Db4k\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 17 Aug 2021 13:15:14 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/AddLineItemSuccess"
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Token Not Found"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "29"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1d-5HbAy3vDHKFvxHIXreOKe2CeD5o\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 17 Aug 2021 13:19:52 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/AddLineItemFailed1"
            }
          }
        },
        "deprecated": false
      }
    },
...
    "/line_items/update/:id": {
      "put": {
        "tags": [
          "line item"
        ],
        "summary": "Update Line Item",
        "operationId": "UpdateLineItem",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "access_token",
            "in": "header",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "qty",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": [
                1
              ]
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "3"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"3-9imuRLez3P7URNNj5ibt9BHsaag\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 17 Aug 2021 13:28:16 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Token Not Found"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "29"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1d-5HbAy3vDHKFvxHIXreOKe2CeD5o\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 17 Aug 2021 13:28:30 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/UpdateLineItemFailed1"
            }
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
...
    "updateimagessuccess": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "updateimagessuccess",
      "example": {
        "message": "Status Has Been Update"
      }
    },
    "updateimagesfail1": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "updateimagesfail1",
      "example": {
        "message": "Token Not Found"
      }
    },
    "deleteimagesfail1": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "deleteimagesfail1",
      "example": {
        "message": "Token Not Found"
      }
    },
...
    "AddLineItemSuccess": {
      "type": "object",
      "required": [
        "OrderId",
        "ProductId",
        "ShoppingCartId",
        "createdAt",
        "id",
        "qty",
        "status",
        "updatedAt"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "qty": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "string"
        },
        "ProductId": {
          "type": "string"
        },
        "ShoppingCartId": {
          "type": "string"
        },
        "OrderId": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        }
      },
      "title": "AddLineItemSuccess",
      "example": {
        "id": 1,
        "qty": 2,
        "status": "Close",
        "updatedAt": "2021-08-17T13:15:14.657Z",
        "createdAt": "2021-08-17T13:15:14.657Z"
      }
    },
    "AddLineItemFailed1": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "AddLineItemFailed1",
      "example": {
        "message": "Token Not Found"
      }
    },
...
    "UpdateLineItemFailed1": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "UpdateLineItemFailed1",
      "example": {
        "message": "Token Not Found"
      }
    }
  }
}